test: factor supervise out of sync helper
